refactor to utilise tel-file picker
