+++ lib/pymedphys/_streamlit/utilities/monaco.py
@@ -2,11 +2,77 @@
 # pylint: disable = no-value-for-parameter, expression-not-assigned
 # pylint: disable = too-many-lines, redefined-outer-name
 
+import os
 import pathlib
 
 from pymedphys._imports import streamlit as st
 
+from pymedphys._monaco import patient as mnc_patient
+from pymedphys._streamlit.utilities import exceptions as _exceptions
+
 from . import exceptions, misc
+
+
+def monaco_tel_files_picker(
+    config, patient_id="", key_namespace="", advanced_mode=False, site=None
+):
+    (
+        monaco_site,
+        monaco_directory,
+        patient_id,
+        plan_directory,
+        patient_directory,
+    ) = monaco_patient_directory_picker(
+        config, patient_id, key_namespace, advanced_mode, site
+    )
+
+    patient_name = read_monaco_patient_name(str(patient_directory))
+
+    st.write(f"Patient Name: `{patient_name}`")
+
+    all_tel_paths = list(plan_directory.glob("**/*tel.1"))
+    all_tel_paths = sorted(all_tel_paths, key=os.path.getmtime, reverse=True)
+
+    plan_names_to_choose_from = [
+        str(path.relative_to(plan_directory)) for path in all_tel_paths
+    ]
+
+    if len(plan_names_to_choose_from) == 0:
+        if patient_id != "":
+            st.write(
+                _exceptions.NoRecordsFound(
+                    f"No Monaco plans found for patient ID {patient_id}"
+                )
+            )
+        return {"patient_id": patient_id}
+
+    st.write(
+        """
+        Select the Monaco plan that correspond to a patient's single fraction.
+        If a patient has multiple fraction types (such as a plan with a boost)
+        then these fraction types need to be analysed separately.
+        """
+    )
+
+    selected_monaco_plan = st.radio(
+        "Select a Monaco plan",
+        plan_names_to_choose_from,
+        key=f"{key_namespace}_monaco_plans",
+    )
+
+    tel_paths = []
+
+    if selected_monaco_plan is not None:
+        current_plans = list(
+            monaco_directory.glob(f"*~{patient_id}/plan/{selected_monaco_plan}")
+        )
+        current_plans = [path.resolve() for path in current_plans]
+        if len(current_plans) != 1:
+            st.write("Plans found:", current_plans)
+            raise ValueError("Exactly one plan should have been found")
+        tel_paths += current_plans
+
+    return monaco_site, monaco_directory, patient_name, selected_monaco_plan, tel_paths
 
 
 def monaco_patient_directory_picker(
@@ -55,3 +121,8 @@ def monaco_patient_directory_picker(
     patient_directory = pathlib.Path(plan_directory).parent
 
     return monaco_site, monaco_directory, patient_id, plan_directory, patient_directory
+
+
+@st.cache
+def read_monaco_patient_name(monaco_patient_directory):
+    return mnc_patient.read_patient_name(monaco_patient_directory)
